+++ database/migrations/2019_09_07_210642_dance_companies_table.php
@@ -14,8 +14,8 @@ class DanceCompaniesTable extends Migration
     public function up()
     {
         Schema::create('dance_companies', function (Blueprint $table) {
-            $table->bigIncrements('id')->unsigned();
-            $table->bigInteger('user_id')->unsigned()->nullable();
+            $table->increments('id')->unsigned();
+            $table->integer('user_id')->unsigned()->nullable();
             $table->string('unique_id', 32)->unique();
             $table->string('name', 96);
             $table->string('city', 96)->nullable();
