+++ database/migrations/2020_01_05_105106_event_attempts.php
@@ -15,7 +15,7 @@ class EventAttempts extends Migration
     {
         Schema::create('dance_event_attempts', function (Blueprint $table) {
             $table->bigInteger('event_id')->unsigned();
-            $table->bigInteger('company_user_id')->unsigned();
+            $table->integer('company_user_id')->unsigned();
             $table->boolean('visited')->default(false);
             $table->timestamps();
             $table->primary(array('event_id', 'company_user_id'));
